Aims now compiles with options -W -Wall
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/trunk@82692 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsalgo/src/aimsalgo/signalfilter/convolutionfilter_d.h
@@ -244,7 +244,7 @@ namespace aims {
   {
     _dir.assign( 4, true );
     _dir[3] = false;
-    for( int i = 0; i < dir.size() && i < _dir.size(); ++i )
+    for( size_t i = 0; (i < dir.size()) && (i < _dir.size()); ++i )
       _dir[i] = dir[i];
   }
